[WIP] frontend and game
