+++ frontend.py
@@ -2,30 +2,31 @@
 import pyglet
 import backend
 
-map_name = "./test_1.json"
-data = backend.get_data(map_name)
-
-def get_windowsize(data):#in backend?
+'''making of window based on function
+def get_windowsize(data):
     map_field_height = data['layers'][0]['height']
     map_field_width = data['layers'][0]['width']
     tile_height = data['tileheight']
     tile_width = data['tilewidth']
     width = tile_width*map_field_width
     height = tile_height*map_field_height
-    return width,height, tile_width, tile_height
+    return width,height, tile_width, tile_height'''
 
-size = get_windowsize(data)
-width = size[0]
-height = size[1]
-tile_width = size[2]
-tile_height= size[3]
-coordinates = backend.get_coordinates(data)
-tilelist = backend.get_tiles(data)
-state = backend.get_coordinate_dict(coordinates,tilelist)
-real_images = backend.get_real_ids(data)
-#lines 5 to 25 should be in backend/game.py?
+def init_window(data):
+    size = backend.get_windowsize(data)
+    width = size[0]
+    height = size[1]
+    tile_width = size[2]
+    tile_height= size[3]
+    window = pyglet.window.Window(width, height)
+    return window
 
-def draw_board():
+#list of images
+def load_images(data, state):
+    real_images = backend.get_real_ids(data)
+    size = backend.get_windowsize(data)
+    tile_width = size[2]
+    tile_height= size[3]
     images = []
     for key in state:
         if state[key] in real_images:
@@ -39,17 +40,11 @@ def draw_board():
             images.append(img)
     return images
 
-window = pyglet.window.Window(width, height)
-tiles = draw_board()
+#drawing function
+def draw_board(state, images):
 
-#rest should be in game.py???  line 46-54
-@window.event
-def on_draw():
-    window.clear()
-
-    for tile in tiles:
+    for tile in images:
         tile.draw()
 
-pyglet.app.run()
 
 #WIP
